+++ core/src/main/java/com/sonicgdx/sonicswirl/TileMap.java
@@ -36,19 +36,19 @@ public enum TileMap {
     //TODO test class - check if all these are 16 in length
 
     // Classes are reference types so modifying a value would affect all the tiles that are the same.
-    private final byte[] zero = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0};
-    private final byte[] slope = {1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16};
-    private final byte[] full = {16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16};
-    private final byte[] halfh = {8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8}; private final byte[] halfw = {0,0,0,0,0,0,0,0,16,16,16,16,16,16,16,16};
-    private final byte[] rvSlope = {16,15,14,13,12,11,10,9,8,7,6,5,4,3,2,1};
-    private final byte[] tall1 = {1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1};
-    private final byte[] testh = {0,0,1,2,2,3,4,5,5,6,6,7,8,9,9,9}, testw = {0,0,0,0,0,0,0,3,4,5,7,9,10,11,13,14};
+    private final int[] zero = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0};
+    private final int[] slope = {1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16};
+    private final int[] full = {16,16,16,16,16,16,16,16,16,16,16,16,16,16,16,16};
+    private final int[] halfh = {8,8,8,8,8,8,8,8,8,8,8,8,8,8,8,8}; private final int[] halfw = {0,0,0,0,0,0,0,0,16,16,16,16,16,16,16,16};
+    private final int[] rvSlope = {16,15,14,13,12,11,10,9,8,7,6,5,4,3,2,1};
+    private final int[] tall1 = {1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1};
+    private final int[] testh = {0,0,1,2,2,3,4,5,5,6,6,7,8,9,9,9}, testw = {0,0,0,0,0,0,0,3,4,5,7,9,10,11,13,14};
     private final Tile EMPTY = new Tile();
-    private final Tile ftile = new Tile(full,full,0,(byte) 4,false,false);
-    private final Tile stile = new Tile(slope, slope,45,(byte) 1,false,false);
-    private final Tile rvtile = new Tile(rvSlope, slope,-45,(byte) 1,true,false);
-    private final Tile htile = new Tile(halfh,halfw,0,(byte) 1,false,false);
-    private final Tile testtile = new Tile(testh,testw,33.75F,(byte) 1,false,false);
+    private final Tile ftile = new Tile(full,full,0,4,false,false);
+    private final Tile stile = new Tile(slope, slope,45,1,false,false);
+    private final Tile rvtile = new Tile(rvSlope, slope,-45,1,true,false);
+    private final Tile htile = new Tile(halfh,halfw,0,1,false,false);
+    private final Tile testtile = new Tile(testh,testw,33.75F,1,false,false);
 
     private final Chunk fChunk = new Chunk(new Texture(Gdx.files.internal("sprites/AIZ2/95.png")),Collections.nCopies(6,Collections.nCopies(6,ftile).toArray(new Tile[0])).toArray(new Tile[0][0]));
 
@@ -134,14 +134,14 @@ public enum TileMap {
     }
 
     /*@Deprecated
-    public byte getHeight(int chunkX, int chunkY, int tileX, int tileY, int block)
+    public int getHeight(int chunkX, int chunkY, int tileX, int tileY, int block)
     {
         if (map[chunkX][chunkY].getTileArray()[tileX][tileY].empty) return 0;
         else return map[chunkX][chunkY].getTileArray()[tileX][tileY].heightArray[block];
     }
 
     @Deprecated
-    public byte getWidth(int chunkX, int chunkY, int tileX, int tileY, int block)
+    public int getWidth(int chunkX, int chunkY, int tileX, int tileY, int block)
     {
         if (map[chunkX][chunkY].getTileArray()[tileX][tileY].empty) return 0;
         else return map[chunkX][chunkY].getTileArray()[tileX][tileY].widthArray[block];
@@ -153,7 +153,7 @@ public enum TileMap {
     }
 
     @Deprecated
-    public byte getHeightAbove(int chunkX, int chunkY, int tileX, int tileY, int block)
+    public int getHeightAbove(int chunkX, int chunkY, int tileX, int tileY, int block)
     {
         if (tileY < 7) tileY = tileY + 1;
         else
@@ -166,7 +166,7 @@ public enum TileMap {
         else return map[chunkX][chunkY].getTileArray()[tileX][tileY].heightArray[block];
     }
     @Deprecated
-    public byte getHeightBelow(int chunkX, int chunkY, int tileX, int tileY, int block)
+    public int getHeightBelow(int chunkX, int chunkY, int tileX, int tileY, int block)
     {
         if (tileY == 0)
         {
@@ -180,16 +180,16 @@ public enum TileMap {
     }
 
     @Deprecated
-    public byte[] getWidthArray(int chunkX, int chunkY, int tileX, int tileY)
+    public int[] getWidthArray(int chunkX, int chunkY, int tileX, int tileY)
     {
-        if (map[chunkX][chunkY].getTileArray()[tileX][tileY].empty) return new byte[16];
+        if (map[chunkX][chunkY].getTileArray()[tileX][tileY].empty) return new int[16];
         else return map[chunkX][chunkY].getTileArray()[tileX][tileY].widthArray;
     }
 
     @Deprecated
-    public byte[] getHeightArray(int chunkX, int chunkY, int tileX, int tileY)
+    public int[] getHeightArray(int chunkX, int chunkY, int tileX, int tileY)
     {
-        if (map[chunkX][chunkY].getTileArray()[tileX][tileY].empty) return new byte[16];
+        if (map[chunkX][chunkY].getTileArray()[tileX][tileY].empty) return new int[16];
         else return map[chunkX][chunkY].getTileArray()[tileX][tileY].heightArray;
     }
 */
